libtest: Prioritize IP over ICMP
There appears to be a priority in dispatching the "alt".

This is either a bug or undocumented feature.

So, for the time being, put the IP above the ICMP
(see test_tcp_undersized()).

For the time being let's go with this workaround,
but let's properly fix as soon as it's feasible.