port_data: Use a separate socket for reading and writing
The port code is ready to be used with two separate endpoints with different src_ip and dst_ip, but one common UDP socket was used for both of these endpoints. Added a separate fd for reading and writing, so enabling user having a separate source and destination IPs.

Fixed a minor bug in setting the src_ip as a parameter.